unified_log: Fix sqlite filtering incorrect timestamp format
The log entries timestamp is a double, we need to ensure
that it's cast to an integer when returning it as a result
to sqlite and using a timestamp constraint,
otherwise sqlite will filter out all the entries
that have the decimal part in them.

+++ osquery/tables/system/darwin/unified_log.mm
@@ -380,7 +380,8 @@ QueryData genUnifiedLog(QueryContext& queryContext) {
 
         Row r;
 
-        r["timestamp"] = BIGINT([[entry date] timeIntervalSince1970]);
+        r["timestamp"] = BIGINT(
+            static_cast<std::uint32_t>([[entry date] timeIntervalSince1970]));
         r["message"] = SQL_TEXT(
             std::string([[entry composedMessage] UTF8String],
                         [[entry composedMessage]
